Introducing balt function wrappers in Link:
_Implementation detail_ - mark the leading underscore.
I need it to single out the non trivial uses of self.window and self.gTank
(on its way to extinction).
Plus the business code in Execute and co in Links subclasses _needs not
know which (or what) is the parent of the dialog_.

Took the opportunity to wrap a bit and also added names for keyword
arguments in (quite a few) places.

Fixed a regression introduced in 8c0d6a7f69bb39b14b45db8f00edd25e1f1ff8af -

  if self.CBashMismatch:
-            if not balt.askContinue(self.window,
-                    _("...Wrye Bash will attempt to import your settings over...
+            if not balt.askYes(self.window,
+                    _("...Should Wrye Bash attempt to import your settings over...

But the askContinue continueKey parameter was left behind and displayed as the
window title: 'bash.patch.CBashMismatch'

Would be nice to keep this and have a checkbox like:
checkBoxTxt=_(u'Always import the old config.')
Those balt methods should be really classes.